+++ packages/google_maps_flutter/google_maps_flutter_web/lib/src/convert.dart
@@ -228,7 +228,7 @@ gmaps.InfoWindowOptions? _infoWindowOptionsFromMarker(Marker marker) {
 
   // If both the title and snippet of an infowindow are empty, we don't really
   // want an infowindow...
-  if ((markerTitle.isEmpty) && (markerSnippet.isEmpty)) {
+  if (markerTitle.isEmpty && markerSnippet.isEmpty) {
     return null;
   }
 
@@ -275,6 +275,8 @@ gmaps.InfoWindowOptions? _infoWindowOptionsFromMarker(Marker marker) {
 
   return gmaps.InfoWindowOptions()
     ..content = container
+    // The deprecated parameter is used here to avoid losing precision.
+    // ignore: deprecated_member_use
     ..zIndex = marker.zIndex;
   // TODO(ditman): Compute the pixelOffset of the infoWindow, from the size of the Marker,
   // and the marker.infoWindow.anchor property.
@@ -456,8 +458,8 @@ Future<gmaps.Icon?> _gmIconFromBitmapDescriptor(
   return icon;
 }
 
-// Computes the options for a new [gmaps.Marker] from an incoming set of options
-// [marker], and the existing marker registered with the map: [currentMarker].
+/// Computes the options for a new [gmaps.Marker] from an incoming set of options
+/// [marker], and the existing marker registered with the map: [currentMarker].
 Future<gmaps.MarkerOptions> _markerOptionsFromMarker(
   Marker marker,
   gmaps.Marker? currentMarker,
@@ -468,6 +470,8 @@ Future<gmaps.MarkerOptions> _markerOptionsFromMarker(
       marker.position.longitude,
     )
     ..title = sanitizeHtml(marker.infoWindow.title ?? '')
+    // The deprecated parameter is used here to avoid losing precision.
+    // ignore: deprecated_member_use
     ..zIndex = marker.zIndex
     ..visible = marker.visible
     ..opacity = marker.alpha