+++ packages/interactive_media_ads/lib/src/ad_display_container.dart
@@ -4,7 +4,7 @@
 
 import 'package:flutter/cupertino.dart';
 
-import 'platform_interface/platform_ad_display_container.dart';
+import 'companion_ad_slot.dart';
 import 'platform_interface/platform_interface.dart';
 
 /// A [Widget] for displaying loaded ads.
@@ -37,6 +37,7 @@ class AdDisplayContainer extends StatelessWidget {
   AdDisplayContainer({
     Key? key,
     required void Function(AdDisplayContainer container) onContainerAdded,
+    Iterable<CompanionAdSlot> companionSlots = const <CompanionAdSlot>[],
     TextDirection layoutDirection = TextDirection.ltr,
   }) : this.fromPlatformCreationParams(
           key: key,
@@ -46,6 +47,9 @@ class AdDisplayContainer extends StatelessWidget {
                 platform: container,
               ));
             },
+            companionSlots: companionSlots.map(
+              (CompanionAdSlot slot) => slot.platform,
+            ),
             layoutDirection: layoutDirection,
           ),
         );
@@ -96,6 +100,12 @@ class AdDisplayContainer extends StatelessWidget {
   /// platform view hierarchy.
   void Function(PlatformAdDisplayContainer container) get onContainerAdded =>
       platform.params.onContainerAdded;
+
+  /// List of companion ad slots.
+  Iterable<CompanionAdSlot> get companionSlots =>
+      platform.params.companionSlots.map(
+        (PlatformCompanionAdSlot slot) => CompanionAdSlot.fromPlatform(slot),
+      );
 
   /// The layout direction to use for the embedded AdDisplayContainer.
   TextDirection get layoutDirection => platform.params.layoutDirection;