+++ packages/video_player/video_player_platform_interface/CHANGELOG.md
@@ -1,5 +1,6 @@
-## NEXT
+## 6.3.1
 
+* Add html 5 video poster support as a VideoPlayerWebOptions.
 * Updates minimum supported SDK version to Flutter 3.27/Dart 3.6.
 
 ## 6.3.0